backend/requirements_back.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

torch
torchvision
timm==1.0.25

opencv-python
Pillow==11.0.0
albumentations==1.4.20

numpy<2
pydantic==2.10.3